+++ src/helpers/app/index.ts
@@ -1,1 +1,2 @@
+export * from './getIsInWebview';
 export * from './provider';